requirements.txt
requests>=2.28
tqdm
arxiv>=1.4.0
pdfplumber>=0.8.0
pypdf>=3.14.0
openai>=1.0.0
python-dotenv
pandas
sqlalchemy
tiktoken